swupd-client: Add Search Feature
New swupd primary command, offers the ability to
search for the provider of a specified binary or library.

A full complement of bundle manifests for the os-release version
is downloaded to the client prior to conducting any search. This
patch has been designed to enable this download, or staging, to
take place prior and separate from use of the search function.
However, with each search, the manifest complement is checked to
ensure completion.

Overhead:

A check against the Clear Linux 6300 manifest set
shows a search will entail a ONE TIME 83 MB network download
(the compressed manifest set), and a constant decompressed 400MB
usage on disk. Future searches will only require new or modified
manifests be downloaded.

Finally, download size is provided to notify user of expected delay

-----
Ex:
"Downloading manifests. Downloading 83.23 MB..."

Signed-off-by: Brad T. Peters <brad.t.peters@intel.com>