Remove symlink check for a path while staging a file.
The do_staging function checks if a path to a file is present before staging a file.
It does so by checking first if it is a symlink and following the symlink.
Remove this check as the way manifests are generated on the server side, it is not
possible to have a file path listed with symlinks in them. The server checks for file
type DT_DIR before including files in that path.