Fix installation of missing boot files with bundle-add
A user was running into issues installing the 'kernel-container' bundle
on Clear Linux. Namely, all missing files marked as boot were not
installed.

The root issue is that the bundle-add action does not set the "fix"
global variable before calling ignore(), so ignore() mistakenly returns
true when considering boot files that are missing, but not marked
deleted.

Since there are only three callers of ignore(), I have removed the
global variable and instead pass in the required values of "fix" to the
new second argument to ignore, "fix_missing". Passing "true" for the
bundle-add action corrects the return value of ignore() in this case.

Signed-off-by: Patrick McCarty <patrick.mccarty@intel.com>