Read default URLs from installed config files
Instead of hardcoding the default version URL and content URL in the
binary, read the defaults from config files that live in
/usr/share/defaults/swupd/.

If the config files do not exist, then users can pass the -c or -v
options as necessary (or -u, which sets the same value for -c and -v).

The primary motivation for this change is to enable a better experience
for the swupd mixer; users mixing a version of Clear Linux can set the
default URLs at mix time instead of always passing -c/-v/-u at runtime.

Signed-off-by: Patrick McCarty <patrick.mccarty@intel.com>