Fix directory removal for verify --fix
In manifests, deleted files/links are not distinguished from deleted
directories, so we cannot rely on the value file->is_dir to indicate how
the delete should happen. Instead, call S_ISDIR(), since lstat() was run
immediately before.

Also, fix the reporting of a successful directory removal.

Note that there is still a bug with sorting of deleted files in
manifests in that the directories should be deleted only after their
containing files/dirs/etc are deleted, but in the case of a standalone
empty directory needing to be deleted, the current code covers that.

Signed-off-by: Patrick McCarty <patrick.mccarty@intel.com>