add --statedir option to alternate state directory
This option overrides the global setup for state_dir
variable that is now used to find out where swupd
is going to do work (state dir).

If not --statedir option is passed along, the global
STATE_DIR definition value is going to be used instead,
this last one can also be changed at compilation time.

Signed-off-by: Jaime A. Garcia <jaime.garcia.naranjo@intel.com>