Fix manifest consolidation when files are deleted in different versions
In the case where a file is deleted from all bundles in the latest
version, but the bundles may have deleted the files in different
versions, make sure the latest version is always chosen to consider for
the update.

I have encountered updates where a deleted file was selected during
consolidation that was deleted in an older version than the version I
updated from, although it was deleted more recently in another bundle.
Since the older deleted file does not get added to the update list, the
file is not deleted during the update.

To avoid these scenarios, always consider the case when two files are
marked deleted in the sorted list, and choose the newer of the two.
Note that this case must be considered before the other delete
conditions, because it is more specific.

Signed-off-by: Patrick McCarty <patrick.mccarty@intel.com>