staging.c: Protect tar command against special characters
It may happen that a bundle contains a directory named '#' and
other files under this directory, thus not only target files
need to be escaped in tar commands, but also target directories
where the files get installed.

Also a target file may have a name with '@' as its first symbol.
Since the symbol has a special meaning in case of bsdtar the
name needs to escaped in tar commands with the prefix './'.

Signed-off-by: Dmitry Rozhkov <dmitry.rozhkov@linux.intel.com>