Add helper to create a file list from bundles
In order to support verify_fix_path from bundle_add, a generic file list
from bundle list needs to be implemented to replace the one embedded in
the current consolidate_submanifests function.

This function will be used to generate a file list to be passed to a new
function for consolidating a list of files (created in the following
patch).

Note the list of files this creates is not sorted.