Implement bundle manifest hash checks
In order for the chain of trust rooted at the Manifest.MoM to operate
correctly, the bundle manifest hashes (as listed in the MoM) must be
checked. For now, warnings will be emitted when hash checks fail, but
they will become errors in the near future.

This commit simplifies/splits the load_manifests() interface into
load_mom() and load_manifest(), since load_manifests() was becoming too
complex, since the handling needs for MoMs versus bundle manifests is
sufficiently different.  For the new load_manifest(), the logic is
changed to first download the manifest, then check the hash, and only
then load the manifest into memory.

Signed-off-by: Patrick McCarty <patrick.mccarty@intel.com>