Simplify and cleanup validate_signature
We should be more strict about the files we are reading in for this function,
as incorrect lengths will completely throw off the hashing. Since the MoM is
relatively small, we should be able to load it all into memory and pass it all
as a single buffer to the context instead of doing it in chunks, which would
make sense for network transfers or very large files. We also need to be using
the sha512 algorithm to hash the data.

Signed-off-by: Tudor Marcu <tudor.marcu@intel.com>