move libcrypto init/cleanup to swupd_init()/swupd_deinit()
The poorly documented libcrypto in OpenSSL apparently does not like
its initialization and cleanup functions to be called but at process
start and exit.  Swupd-client already has a swupd_init() so easy on
that side.  But there was no common swupd_deinit() in which to place
terminate_signature().  I add one and put the common exit cleaners in it
and fix of a number of little inconsistencies around process exit cleanup
that have come to exist because there was a common cleaner.

With this, the signature verification of both the current and latest
MoM works, where prior only a first verification succeeded.

This patch also fixes a few memory leaks, though there are still many
memory leaks in swupd bundle-add and bundle-remove, unrelated to signature
verification.  And swupd search is segfaulting due to how load_mom()
is introduced in 8cf0ef91ddf737c8ceb92abc664c743d9e3e9349.

Signed-off-by: Tim Pepper <timothy.c.pepper@linux.intel.com>