Release v3.6.0
This release enables support for verification of signed MoM manifests,
with the MoM being the root of trust for any swupd-client action that
references update content (i.e.  'update', 'verify', etc.). Also, to
complete the chain of trust rooted at the MoM, integrity checks have
been enabled for bundle manifests.

To enable this feature, pass the --enable-signature-verification option
to the configure script.

Signed-off-by: Patrick McCarty <patrick.mccarty@intel.com>