Move sign/verify to abort on failure
Move swupd signature verification to mandatory. Failed
signature verification now aborts 'swupd update' and 'swupd verify'.
Their operation may be forced to proceed past the signature
verification if the following flag is passed in:

  --force-continue-with-unverified-signature

The signed MoM serves as the top level chain of trust as
far as swupd is concerned, and it is used to extend content
trust down to the individual file level.

Signed-off-by: Brad T. Peters <brad.t.peters@intel.com>