show version when emitting bundle manifest hash mismatch warning
We've had some issues where bundle manifest hashes were
mismatching against the MoM.  We think we know why, but are
partly guessing, because we don't know which version of bundle
manifest was in play.  We know the version at the point of the
printf, so add it to the output.

Signed-off-by: Tim Pepper <timothy.c.pepper@linux.intel.com>