Do not mark a pack file as "downloaded and extracted" on tar error.
Because this code did not ever look at the return value of
`tar`, it is possible that a download succeeded to download
something (possibly a bad file on the server, etc.) properly,
and tar then failed to extract it (possible local problems
with disk space, a random memory failure, etc.).

When this condition happens, we should never leave an empty
pack file on the system, since we're sure that there was an
error extracting it, and the best course of action is to
either re-download it or download individual files. Of course,
the latter is likely going to happen anyway.