Fix file install failure during update
In some circumstances, files were incorrectly not being installed
during an update.

Currently when a new bundle (A) is added by being included in another
already installed bundle (B) as part of update and bundle A contains a
file (F) not installed in the current version AND B does not include F
in the current version but will include F after the update the file
will erroneously not be installed after the update completes.

The fix for this issue is to check during the manifest link stage of
update if the server and current version contain the same file
(matching hashes and not deleted) but the current version is not
currently tracked (the bundle containing the file isn't installed) to
not create the link between current and server versions. This allows
the tracked server version to be used in consolidate_files instead of
the untracked current version. With the correct version being used,
create_update_list will add the file to the update list because it
will be newer than the current OS version and be added to the
filesystem once the update completes successfully.

Signed-off-by: Brad T. Peters <brad.t.peters@intel.com>