Recover from missing or bad bundle manifest files.
This addresses #94.

We can't verify that bundles are correct if a bundle manifest we need
is missing or corrupted.

We first check if these files are missing, and download them before
verifying them.

If they are present or have just been downloaded, we can verify the
hash checksum. If this fails to verify, we immediately erase the bad
file and return an error code.

Because the overlying routines properly retries, we'll get back to the
start of this routine again and download the proper bundle manifest
file, completing the recovery.