Move signature verification to Mandatory
Moves swupd signature verification to mandatory, though --force
is supported in some cases, described here:

The signed MoM serves as the top level chain
of trust as far as swupd is concerned, and it is used to extend
content trust down to the individual file level. When the signature
of the top-level MoM is invalid or cannot be verified for any reason,
we warn and abort the operation.
  --force will allow this to proceed, explicitly accepting the
	unverified MoM and outputting a log entry to the Journal to this
	effect. This may introduce a security hole

For files being downloaded from a Clear server, the downloaded file
must match the MoM hash, or a warning is emitted and the operation is
aborted.
  --force has no effect

Signed-off-by: Brad T. Peters <brad.t.peters@intel.com>