Fix hysteresis logic for fullfile downloads
The second argument to curl_multi_perform(), populated by libcurl, is
currently being used to account for the remaining transfers in the
multi-stack.  But that usage is incorrect; that value instead indicates
the number of active transfers in progress.

To properly enforce the hysteresis logic, use a counter variable
(mcurl_size) instead to track the multi-stack size.

A few other notes:

- Leverage curl_multi_wait() instead of calling usleep(). The last
  argument of curl_multi_wait() can be inspected after a call to see how
  many "interesting" events occurred. In practice, this number appears to
  be an upper bound, so we can at least try to process that many
  transfers.

- Call curl_multi_perform() only in poll_fewer_than(), to more cleanly
  separate the logic in the file.

Signed-off-by: Patrick McCarty <patrick.mccarty@intel.com>