Make remove() error messages correct and useful.
We do not want to print this message if there was no file to begin
with, so check that errno isn't ENOENT.

We also want to print strerror(errno) instead of a code, and last,
we want to print the right filename here as well.

This is due to commit f06bcdba6d5b994a426cf23c76b97c79989f6427.

We should seriously consider also reverting that commit, as there
is almost no value in these error messages.