Initialize tmprenamedir in the state directory
Initializing this directory here fixes an issue with creating custom
Clear Linux Docker images with swupd; when "tmprenamedir" doesn't exist
in the image layer and is later created in the container layer, a
rename() call will fail with EXDEV, due to the layers existing on
different overlayed filesystems. Ensuring that "tmprenamedir" is always
present in the state directory resolves the issue.

Signed-off-by: Patrick McCarty <patrick.mccarty@intel.com>