update: rework update list creation
The purpose of calling verify_file() in create_update_list() was to
minimize extraneous fullfile downloads when a minversion bump occurs
server-side, and the update crosses that version.

However, the implementation was buggy; verify_file() was called far too
often, namely for any file on the system that did not change. This
commit fixes the issue.

Signed-off-by: Patrick McCarty <patrick.mccarty@intel.com>