tar: extract failure handling
tar errors when unpacking individual files used to be ignored in swupd
client < 3.0 and current 3.6.0 still prints a message about "ignoring
tar extract failure for fullfile ..." but the error is not really
getting ignored. Instead, downloading is tried again a few times
before swupd gives up.

This is relevant in Ostro OS where IMA prevents bsdtar from setting a
security.ima xattr when the value is redundant (see "[Linux-ima-user]
copying files with security.ima xattr" - the consensus is that the
kernel is too strict and needs to be fixed), leading to errors that
can be ignored.

The same can happen when extracting a pack file, and there errors also
should be ignored.

In both cases it is useful to see the actual error messages from the
tar command, so the redirection to /dev/null gets removed.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>