Add man (manual) pages.
Add a large amount of manual pages to swupd-client project. These are
meant to cover a large amount of topics and should cover the needed
material for people who are either using, or mixing with swupd,
and so this goes a bit further than end user documentation.

I've put the unit stuff in section 4. Any other section seemed
less related.