staging: better handle file type change from directory
If the file type changes from directory to something else,
we have to make sure to end up in the right place when
extracting it.

So far, we end up the directory case, but of course we
need to get into the case of non-directory.

Note that at least bsdtar handles extracting a new non-
directory over an existing directory with the same name
fine, i.e. the existing directory is correctly removed,
and the new file (or symlink) created instead.

Signed-off-by: André Draszik <git@andred.net>