verify: if downloads fail, do not report stats
When pack or fullfiles fail to download for 'verify --install' or
'verify --fix', respectively, the operation fails, but it does not leave
the system in a corrupt state, so there is no need to report stats in
this situation.

Also, make sure to use errors from the download functions in case
they return specific error codes (converted here to their positive
variants).

Signed-off-by: Patrick McCarty <patrick.mccarty@intel.com>