Avoid cleaning the multi queue too early
There may be in-progress transfers in the multi queue if
poll_fewer_than() returns an error, so a full clean of the multi queue
should be deferred.

The queue is already cleaned up immediately before the next download
retry loop begins; this is the best place for cleanup, since all
transfers have either completed or failed before the cleanup is
triggered.

Signed-off-by: Patrick McCarty <patrick.mccarty@intel.com>